Removed some Type API in Identity/Entity that was now senseless
